Clarity: trip registrations end on "Event end date"
This commit is targeted at our "fake trips", such as ISC hoodie or
Revolut cards trip. The purpose of these fake trips is usually to
collect money or track something. This action is going on during the
trip, i.e. from "Event starts date" (Trip::event->datetime_from) to
"Event end date" (Trip::trip_date_to), however, with the current
setting, the trip registrations end on the 7th day after the "Event
starts date" so we're not able to register people after the first 7 days
of the event. That's where this change comes in. The end of registration
period was changed to "Event end date" to match the expected behavior.

Side effect:
Registering people for an ongoing trip is possible. That's not a problem
though, because sometimes people who are not registered come to the trip
anyway. This is a way to register them and have consistent data on paper
and PartakNET.